Fixes: Adjust function name; Add pragma; Add missing includes; Access raw memory and use reinterp_cast in CUDA implementation to integrate existing functions from iou_box3d

+++ point_frustums/ops/nms.py
@@ -20,4 +20,4 @@ def nms_3d(
     if labels.numel() == 0:
         return torch.empty_like(labels).bool()
 
-    return torch.ops.point_frustums.nms(labels, scores, boxes, iou_threshold, distance_threshold).bool()
+    return torch.ops.point_frustums.nms(labels.int(), scores, boxes, iou_threshold, distance_threshold).bool()
